mengganti method parseInt dengan method ParseFloat agar bisa mengkalkulasi decimal

+++ script.js
@@ -54,16 +54,16 @@ const calculate = () => {
     switch (calculationOperator){
 
     case "+":
-        result = parseInt (prevNumber) + parseInt (currentNumber)
+        result = parseFloat (prevNumber) + parseFloat (currentNumber)
         break
     case "-":
-        result = parseInt (prevNumber) - parseInt (currentNumber)
+        result = parseFloat (prevNumber) - parseFloat (currentNumber)
         break
     case "*":
-        result = parseInt (prevNumber) * parseInt (currentNumber)
+        result = parseFloat (prevNumber) * parseFloat (currentNumber)
         break
     case "/":
-        result = parseInt (prevNumber) / parseInt (currentNumber)
+        result = parseFloat (prevNumber) / parseFloat (currentNumber)
         break
     default:
         return
